Crate musli_storage

Super simple storage encoding for Müsli

The storage encoding is partially upgrade safe:

✔ Can tolerate missing fields if they are annotated with #[musli(default)].

✗ Cannot skip over extra unrecognized fields.

This means that it’s suitable as a storage format, since the data model only evolves in one place. But unsuitable as a wire format since it cannot allow clients to upgrade independent of each other.

See musli-wire for a fully upgrade safe format.

use musli::{Encode, Decode};

#[derive(Debug, PartialEq, Encode, Decode)]
struct Version1 {
    name: String,
}

#[derive(Debug, PartialEq, Encode, Decode)]
struct Version2 {
    name: String,
    #[musli(default)]
    age: Option<u32>,
}

let version2 = musli_storage::to_buffer(&Version2 {
    name: String::from("Aristotle"),
    age: Some(62),
})?;

assert!(musli_storage::decode::<_, Version1>(version2.as_slice()).is_err());

let version1 = musli_storage::to_buffer(&Version1 {
    name: String::from("Aristotle"),
})?;

let version2: Version2 = musli_storage::decode(version1.as_slice())?;

assert_eq!(version2, Version2 {
    name: String::from("Aristotle"),
    age: None,
});

Configuring

To tweak the behavior of the storage format you can use the Encoding type:

use musli_storage::Encoding;
use musli_storage::int::{Fixed, Variable};
use musli::mode::DefaultMode;
use musli::{Encode, Decode};

const CONFIG: Encoding<DefaultMode, Fixed, Variable> = Encoding::new()
    .with_fixed_integers();

#[derive(Debug, PartialEq, Encode, Decode)]
struct Struct<'a> {
    name: &'a str,
    age: u32,
}

let mut out = Vec::new();

let expected = Struct {
    name: "Aristotle",
    age: 61,
};

CONFIG.encode(&mut out, &expected)?;
let actual = CONFIG.decode(&out[..])?;

assert_eq!(expected, actual);

Re-exports

pub use self::encoding::decode;

pub use self::encoding::encode;

pub use self::encoding::from_slice;

pub use self::encoding::to_buffer;

pub use self::encoding::to_fixed_bytes;

pub use self::encoding::Encoding;

pub use self::encoding::to_vec;

pub use self::encoding::to_writer;

Modules

buffered_writer

A writer which buffers the writes before it outputs it into the backing storage.

encoding

Module that defines Encoding whith allows for customization of the encoding format, and the DEFAULT encoding configuration.

error

Generic error types that can be used for most Reader / Writer implementations.

fixed_bytes

A container which can store up to a fixed number of uninitialized bytes on the stack and read into and from it.

int

Traits and utilities for dealing with integers.

reader

Trait for governing how a particular source of bytes is read.

wrap

Helpers for integrating musli with I/O types like std::io and std::io::Write.

writer

Trait for governing how a particular sink of bytes is written to.