Module musli_storage::encoding

Module that defines Encoding whith allows for customization of the encoding format, and the DEFAULT encoding configuration.

Structs

Encoding

Setting up encoding with parameters.

Constants

DEFAULT

The default configuration.

Functions

decode

Decode the given type T from the given Reader using the DEFAULT configuration.

encode

Encode the given value to the given Writer using the DEFAULT configuration.

from_slice

Decode the given type T from the given slice using the DEFAULT configuration.

to_buffer

Encode the given value to a Buffer using the DEFAULT configuration.

to_fixed_bytes

Encode the given value to a fixed-size bytes using the DEFAULT configuration.

to_vec

Encode the given value to a Vec using the DEFAULT configuration.

to_writer

Encode the given value to the given Write using the DEFAULT configuration.