Trait musli_storage::int::UsizeEncoding

pub trait UsizeEncoding: Sealed {
    // Required methods
    fn encode_usize<W>(
        writer: W,
        value: usize
    ) -> Result<(), <W as Writer>::Error>
       where W: Writer;
    fn decode_usize<'de, R>(
        reader: R
    ) -> Result<usize, <R as Reader<'de>>::Error>
       where R: Reader<'de>;
}

Encoding formats which ensure that variably sized types (like usize, isize) are encoded in a format which is platform-neutral.

Required Methods

fn encode_usize<W>(writer: W, value: usize) -> Result<(), <W as Writer>::Error>
where W: Writer,

Governs how usize lengths are encoded into a Writer.

fn decode_usize<'de, R>(reader: R) -> Result<usize, <R as Reader<'de>>::Error>
where R: Reader<'de>,

Governs how usize lengths are decoded from a Reader.

Object Safety

This trait is not object safe.