Module musli_storage::reader

Trait for governing how a particular source of bytes is read.

musli requires all sources to reference the complete data being read from it which allows it to make the assumption the bytes are always returned with the 'de lifetime.

Structs

Limit

Limit the number of bytes that can be read out of a reader to the specified limit.

SliceReader

An efficient Reader wrapper around a slice.

WithPosition

Keep a record of the current position.

Traits

PosReader

A reader where the current position is exactly known.

Reader

Trait governing how a source of bytes is read.