Trait musli_storage::reader::Reader

pub trait Reader<'de> {
    type Error: Error;
    type Mut<'this>: Reader<'de, Error = Self::Error>
       where Self: 'this;

    // Required methods
    fn borrow_mut(&mut self) -> Self::Mut<'_>;
    fn skip(&mut self, n: usize) -> Result<(), Self::Error>;
    fn peek(&mut self) -> Result<Option<u8>, Self::Error>;
    fn read_bytes<V>(
        &mut self,
        n: usize,
        visitor: V
    ) -> Result<<V as ValueVisitor<'de>>::Ok, <V as ValueVisitor<'de>>::Error>
       where V: ValueVisitor<'de, Target = [u8], Error = Self::Error>;

    // Provided methods
    fn read(&mut self, buf: &mut [u8]) -> Result<(), Self::Error> { ... }
    fn read_byte(&mut self) -> Result<u8, Self::Error> { ... }
    fn read_array<const N: usize>(&mut self) -> Result<[u8; N], Self::Error> { ... }
    fn with_position(self) -> WithPosition<Self>
       where Self: Sized { ... }
    fn limit(self, limit: usize) -> Limit<Self>
       where Self: Sized { ... }
}

Trait governing how a source of bytes is read.

This requires the reader to be able to hand out contiguous references to the byte source through Reader::read_bytes.

Required Associated Types

type Error: Error

Error type raised by the current reader.

type Mut<'this>: Reader<'de, Error = Self::Error> where Self: 'this

Reborrowed type.

Why oh why would we want to do this over having a simple &'this mut T?

We want to avoid recursive types, which will blow up the compiler. And the above is a typical example of when that can go wrong. This ensures that each call to borrow_mut dereferences the Reader at each step to avoid constructing a large muted type, like &mut &mut &mut SliceReader<'de>.

Required Methods

fn borrow_mut(&mut self) -> Self::Mut<'_>

Reborrow the current reader.

fn skip(&mut self, n: usize) -> Result<(), Self::Error>

Skip over the given number of bytes.

fn peek(&mut self) -> Result<Option<u8>, Self::Error>

Peek the next value.

fn read_bytes<V>( &mut self, n: usize, visitor: V ) -> Result<<V as ValueVisitor<'de>>::Ok, <V as ValueVisitor<'de>>::Error>
where V: ValueVisitor<'de, Target = [u8], Error = Self::Error>,

Read a slice out of the current reader.

Provided Methods

fn read(&mut self, buf: &mut [u8]) -> Result<(), Self::Error>

Read a slice into the given buffer.

fn read_byte(&mut self) -> Result<u8, Self::Error>

Read a single byte.

fn read_array<const N: usize>(&mut self) -> Result<[u8; N], Self::Error>

Read an array out of the current reader.

fn with_position(self) -> WithPosition<Self>
where Self: Sized,

Keep an accurate record of the position within the reader.

fn limit(self, limit: usize) -> Limit<Self>
where Self: Sized,

Keep an accurate record of the position within the reader.

Object Safety

This trait is not object safe.