Module musli_storage::writer

Trait for governing how a particular sink of bytes is written to.

To adapt std::io::Write types, see the [wrap][crate::io::wrap] function.

Structs

Buffer

A buffer that roughly corresponds to a vector. For no-std environments this has a fixed size and will error in case the size overflows.

Constants

MAX_FIXED_BYTES_LEN

Maximum size used by a fixed length Buffer.

Traits

Writer

The trait governing how a writer works.